chore: move MLS api implementation to v4
