+++ network/src/commonMain/kotlin/com/wire/kalium/network/api/v3/authenticated/KeyPackageApiV3.kt
@@ -18,9 +18,6 @@
 
 package com.wire.kalium.network.api.v3.authenticated
 
-import com.wire.kalium.network.AuthenticatedNetworkClient
 import com.wire.kalium.network.api.v2.authenticated.KeyPackageApiV2
 
-internal open class KeyPackageApiV3 internal constructor(
-    authenticatedNetworkClient: AuthenticatedNetworkClient
-) : KeyPackageApiV2(authenticatedNetworkClient)
+internal open class KeyPackageApiV3 internal constructor() : KeyPackageApiV2()
