+++ network/src/commonMain/kotlin/com/wire/kalium/network/api/v4/authenticated/MLSMessageApiV4.kt
@@ -22,6 +22,7 @@ import com.wire.kalium.network.AuthenticatedNetworkClient
 import com.wire.kalium.network.api.base.authenticated.message.MLSMessageApi
 import com.wire.kalium.network.api.base.authenticated.message.SendMLSMessageResponse
 import com.wire.kalium.network.api.v3.authenticated.MLSMessageApiV3
+import com.wire.kalium.network.serialization.Mls
 import com.wire.kalium.network.serialization.XProtoBuf
 import com.wire.kalium.network.utils.NetworkResponse
 import com.wire.kalium.network.utils.wrapKaliumResponse
@@ -32,9 +33,25 @@ import io.ktor.http.contentType
 
 internal open class MLSMessageApiV4 internal constructor(
     private val authenticatedNetworkClient: AuthenticatedNetworkClient
-) : MLSMessageApiV3(authenticatedNetworkClient) {
+) : MLSMessageApiV3() {
 
     private val httpClient get() = authenticatedNetworkClient.httpClient
+
+    override suspend fun sendMessage(message: MLSMessageApi.Message): NetworkResponse<SendMLSMessageResponse> =
+        wrapKaliumResponse {
+            httpClient.post(PATH_MESSAGE) {
+                setBody(message.value)
+                contentType(ContentType.Message.Mls)
+            }
+        }
+
+    override suspend fun sendWelcomeMessage(message: MLSMessageApi.WelcomeMessage): NetworkResponse<Unit> =
+        wrapKaliumResponse {
+            httpClient.post(PATH_WELCOME_MESSAGE) {
+                contentType(ContentType.Message.Mls)
+                setBody(message.value)
+            }
+        }
 
     override suspend fun sendCommitBundle(bundle: MLSMessageApi.CommitBundle): NetworkResponse<SendMLSMessageResponse> =
         wrapKaliumResponse {
@@ -46,5 +63,7 @@ internal open class MLSMessageApiV4 internal constructor(
 
     private companion object {
         const val PATH_COMMIT_BUNDLES = "mls/commit-bundles"
+        const val PATH_MESSAGE = "mls/messages"
+        const val PATH_WELCOME_MESSAGE = "mls/welcome"
     }
 }
