+++ network/src/commonMain/kotlin/com/wire/kalium/network/api/v4/authenticated/NotificationApiV4.kt
@@ -20,32 +20,11 @@ package com.wire.kalium.network.api.v4.authenticated
 
 import com.wire.kalium.network.AuthenticatedNetworkClient
 import com.wire.kalium.network.AuthenticatedWebSocketClient
-import com.wire.kalium.network.api.base.authenticated.notification.NotificationResponse
 import com.wire.kalium.network.api.v3.authenticated.NotificationApiV3
 import com.wire.kalium.network.tools.ServerConfigDTO
-import com.wire.kalium.network.utils.NetworkResponse
-import com.wire.kalium.network.utils.wrapKaliumResponse
-import io.ktor.client.request.get
-import io.ktor.client.request.parameter
 
 internal open class NotificationApiV4 internal constructor(
     private val authenticatedNetworkClient: AuthenticatedNetworkClient,
     authenticatedWebSocketClient: AuthenticatedWebSocketClient,
     serverLinks: ServerConfigDTO.Links
-) : NotificationApiV3(authenticatedNetworkClient, authenticatedWebSocketClient, serverLinks) {
-
-    private val httpClient get() = authenticatedNetworkClient.httpClient
-
-    override suspend fun notificationsCall(
-        querySize: Int,
-        queryClient: String,
-        querySince: String?
-    ): NetworkResponse<NotificationResponse> = wrapKaliumResponse {
-        // Pretty much the same V0 request, but without the 404 overwrite
-        httpClient.get(V0.PATH_NOTIFICATIONS) {
-            parameter(V0.SIZE_QUERY_KEY, querySize)
-            parameter(V0.CLIENT_QUERY_KEY, queryClient)
-            querySince?.let { parameter(V0.SINCE_QUERY_KEY, it) }
-        }
-    }
-}
+) : NotificationApiV3(authenticatedNetworkClient, authenticatedWebSocketClient, serverLinks)
