--- network/src/commonTest/kotlin/com/wire/kalium/api/v3/MLSMessageApiV3Test.kt
@@ -1,60 +1,0 @@
-/*
- * Wire
- * Copyright (C) 2023 Wire Swiss GmbH
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program. If not, see http://www.gnu.org/licenses/.
- */
-
-package com.wire.kalium.api.v3
-
-import com.wire.kalium.api.ApiTest
-import com.wire.kalium.model.SendMLSMessageResponseJson
-import com.wire.kalium.network.api.base.authenticated.message.MLSMessageApi
-import com.wire.kalium.network.api.v3.authenticated.MLSMessageApiV3
-import com.wire.kalium.network.serialization.XProtoBuf
-import com.wire.kalium.network.utils.isSuccessful
-import io.ktor.http.ContentType
-import io.ktor.http.HttpStatusCode
-import kotlinx.coroutines.ExperimentalCoroutinesApi
-import kotlinx.coroutines.test.runTest
-import kotlin.test.Test
-import kotlin.test.assertTrue
-
-@ExperimentalCoroutinesApi
-class MLSMessageApiV3Test : ApiTest {
-
-    @Test
-    fun givenMessage_whenSendingMessage_theRequestShouldBeConfiguredCorrectly() =
-        runTest {
-            val networkClient = mockAuthenticatedNetworkClient(
-                SendMLSMessageResponseJson.validMessageSentJson.rawJson,
-                statusCode = HttpStatusCode.Created,
-                assertion =
-                {
-                    assertPost()
-                    assertContentType(ContentType.Application.XProtoBuf)
-                    assertPathEqual(PATH_COMMIT_BUNDLES)
-                }
-            )
-            val mlsMessageApi: MLSMessageApi = MLSMessageApiV3(networkClient)
-            val response = mlsMessageApi.sendCommitBundle(COMMIT_BUNDLE)
-            assertTrue(response.isSuccessful())
-        }
-
-    private companion object {
-        const val PATH_COMMIT_BUNDLES = "mls/commit-bundles"
-        val COMMIT_BUNDLE = MLSMessageApi.CommitBundle("CommitBundle".encodeToByteArray())
-    }
-
-}
